+++ projects/swimlane/ngx-cron/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## HEAD (unreleased)
 
+- Breaking: require new version of `@swimlane/ngx-ui`
+- Feature: show timezone popup when timezone is supplied
+- Bug fix: fix timezone bug
+
 ## 3.0.0 (2021-09-14)
 
 - Breaking: Upgrade to Angular 12 (#34)